+++ HULK-Intrepreter/Code Analysis/Syntax/Lexer.cs
@@ -175,6 +175,7 @@ namespace HULK.CodeAnalysis.Syntax
                     }    
                     default:
                         _diagnostics.ReportBadCharacter(_position, Current);
+                        // _kind = SyntaxKind.BadToken
                         _position++;
                         break;
                 }
@@ -202,7 +203,7 @@ namespace HULK.CodeAnalysis.Syntax
             var length = _position - _start;
             var text = _text.Substring(_start, length);
 
-            if (!int.TryParse(text, out var value))
+            if (!double.TryParse(text, out var value))
                 _diagnostics.ReportInvalidNumber(new TextSpan(_start, length), text, typeof(int));
 
             _value = value;
